Handle CentOS 6/7 differences when provisioning

+++ provision.sh
@@ -1,38 +1,62 @@
 #!/bin/bash
 
 if [ ! $UID -eq 0 ]; then
-    echo "You must be root to execute this script!"
-    exit 1
+	echo "You must be root to execute this script!"
+	exit 1
 fi
 
-yum -y install git screen wget
-yum -y groupinstall "Development tools"
-yum -y install zlib-devel bzip2-devel openssl-devel ncurses-devel sqlite-devel readline-devel tk-devel atlas-devel
+if [ -f /etc/centos-release ]; then
+	CENTOS_VERSION=$(rpm -q --queryformat '%{VERSION}' centos-release)
 
-cd $(mktemp -d)
-wget http://python.org/ftp/python/2.7.8/Python-2.7.8.tgz
-tar xf Python-2.7.8.tgz
-cd Python-2.7.8
-./configure --prefix=/usr/local
-make && make altinstall
+	yum -y install git screen wget curl
+	yum -y groupinstall "Development tools"
 
-wget https://pypi.python.org/packages/source/d/distribute/distribute-0.6.49.tar.gz
-tar xf distribute-0.6.49.tar.gz
-cd distribute-0.6.49
-/usr/local/bin/python2.7 setup.py install
-/usr/local/bin/easy_install-2.7 pip
+	if [ $CENTOS_VERSION -eq 6 ]; then
+		yum -y install \
+			zlib-devel \
+			bzip2-devel \
+			openssl-devel \
+			ncurses-devel \
+			sqlite-devel \
+			readline-devel \
+			tk-devel \
+			atlas-devel
 
-cat >/etc/yum.repos.d/docker.repo <<-EOF
-[dockerrepo]
-name=Docker Repository
-baseurl=https://yum.dockerproject.org/repo/main/centos/6
-enabled=1
-gpgcheck=1
-gpgkey=https://yum.dockerproject.org/gpg
-EOF
+		cd $(mktemp -d)
 
-yum -y check-update
-yum -y install docker-engine
+		# XXX Update the starting point with the current version
+		LATEST_PYTHON=10
 
-/usr/local/bin/pip2.7 install docker-compose
-mkdir -p /etc/docker
+		for x in $(seq ${LATEST_PYTHON} 100); do
+			head_req=$(curl -I --silent https://www.python.org/ftp/python/2.7.$x/Python-2.7.$x.tgz)
+			if grep --quiet --perl-regexp  "^HTTP/1\.1 200 OK" <(echo $head_req); then
+				LATEST_PYTHON=$x
+			else
+				break
+			fi
+		done
+
+		wget http://python.org/ftp/python/2.7.${LATEST_PYTHON}/Python-2.7.${LATEST_PYTHON}.tgz
+		tar xf Python-2.7.8.tgz
+		cd Python-2.7.8
+		./configure --prefix=/usr/local
+		make && make altinstall
+	fi
+
+	curl "https://bootstrap.pypa.io/get-pip.py" | python
+
+	cat >/etc/yum.repos.d/docker.repo <<-EOF
+	[dockerrepo]
+	name=Docker Repository
+	baseurl=https://yum.dockerproject.org/repo/main/centos/$CENTOS_VERSION
+	enabled=1
+	gpgcheck=1
+	gpgkey=https://yum.dockerproject.org/gpg
+	EOF
+
+	yum -y check-update
+	yum -y install docker-engine
+	service docker start
+
+	pip install docker-compose
+fi
